Add files via upload
The Knapsack Problem is a classic optimization problem in computer science and mathematics. In its basic form, we are given a knapsack with a fixed capacity (or budget) and a set of items, each with a specific value and weight (or cost). The goal is to select a subset of items such that the total value is maximized without exceeding the knapsack's capacity (budget).